+++ optimisation.c
@@ -6,13 +6,13 @@
 /*   By: rnovotny <rnovotny@student.42prague.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/03 11:43:29 by rnovotny          #+#    #+#             */
-/*   Updated: 2023/10/12 16:49:45 by rnovotny         ###   ########.fr       */
+/*   Updated: 2023/10/16 15:42:31 by rnovotny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	operation_sums(unsigned short *operations, int i, int sum)
+int	operation_sums(int *operations, int i, int sum)
 {
 	int	j;
 
@@ -32,7 +32,7 @@ int	operation_sums(unsigned short *operations, int i, int sum)
 }
 
 
-void	optimise_operations(unsigned short *operations)
+void	optimise_operations(int *operations)
 {
 	int	i;
 	int	j;
@@ -85,7 +85,7 @@ void	write_operation(short i)
 		write(1, "rrr\n", 4);
 }
 
-void	read_operations(unsigned short *operations)
+void	read_operations(int *operations)
 {
 	int	i;
